src/binbotspawner/requirements.txt
functions-framework==3.*
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
twilio==9.0.2
urllib3==2.2.1
google-cloud-logging==3.10.0
google-cloud-firestore==2.16.0
aiohttp==3.9.3